simplified checkFileSets
Originally, checkFileSets attempted to repair symbolic links and other repairs to the file sets.

Improper modifications of package file sets sometimes occurred, which could result in corruption of the active files so the entire mechanism was abandoned.

checkFileSets now only attempts to create a file set for a new Venus OS version.